Add CSS for 'See Results' button and update corresponding HTML
This commit adds CSS for a new 'See Results' button and updates the corresponding HTML to use the new button instead of the previous "See Facts" button. The new button has a distinct style and a hover effect. The input radio elements are also wrapped in a separate div for better structure.